Update VehicleMovementSystem and GraphGeneratorSystem
Now vehicles follow virtual tracks, but the algorithm needs
two physical tracks to work, since one is not enough in all
scenarios.

Graph generation needs a choice: either a track in a cross
exists if and only if the ending street exists and it is not
one-way in the opposite direction; or a track may not exist
even if the ending street exists and it is not one-way or
it is one-way in the current direction.